Create Analysis of Linear Regression
In this, We do the analysis of Linear Regression Using the dummy data Which is actually the Boston data that is present in the sklearn Data set . These data are clean, So there is no need of cleaning the data. Here we use the Linear Regression algorithm which is present initially in the sklearn  module.