MultinomialNB vs sefl implementation
In this part, we are first to clean the data of 20newsgroups and extract the 2d array from it. Then we observe the accuracy by the use of inbuilt Multinomial Naive_bayes and then we compare it with the self-implemented part and from both the results we observe that they give very good results.